+++ doc/gb-model/release_notes.rst
@@ -13,6 +13,7 @@ Upcoming Release
 ================
 
 * Add interconnectors to network.
+* Add boundary capability constraints to GB model (#131).
 * Merge Shetland (region 30) and Northern Ireland (region 31) to other regions (#117).
 * Add demands to pypsa Network (#102, #70, #120).
 * Limit GB model to ``clustered`` clusters.